+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Collection: view settings allow changing the sort order (aka ascending/descending)
 - Collection / Info: edit title via IPTC / XMP
 - Albums / Countries / Tags: size displayed in list view details, sort by size
 - Search: `undated` and `untitled` filters